Huge commit, details below.
This is all part of implementing a config format that can provide a
clean way to describe your SLAs, what they apply to, and organize report
output. The config file specifies Targets, which are a set of Issues
coming from a board or filter, and a set of Rules which simply map an
Issue to a boolean.

Targets have a Name member, which will be used in the output later.

I've generalized the issue fetching. It accepts a Target value ( which
contains board ID or filter ID or both ) and an AuthOptions value for
obvious reasons. This is the simplest way and avoids coupling
configuration stuff to anything else. Actually, AuthOptions is going to
be purged in the next commit so ignore that.

Comitting the compliance code for completeness.